Sistema_gestion_actividades/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.1
django-cors-headers==4.2.0
django-crontab==0.7.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.4
Jinja2==3.1.2
jsonfield==3.1.0
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
openpyxl==3.1.2
packaging==23.1
Pillow==10.0.0
protobuf==3.20.3
PyJWT==2.8.0
PyMySQL==1.0.3
pytz==2023.3
redmail==0.6.0
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.6.2
tzdata==2023.3
urllib3==2.0.4
